requirements.txt
asgiref==3.4.1
Django==3.2.9
django-phone-field==1.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
numpy==1.21.2
pandas==1.3.3
pi==0.1.2
Pillow==8.3.2
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2